+++ Common/Maps/PathFinding/WeightedNode.cs
@@ -2,15 +2,15 @@
 {
     public class WeightedNode : BaseNode
     {
-        public WeightedNode(int weight, Position position) : base(position)
+        public WeightedNode(long weight, Position position) : base(position)
         {
             Weight = weight;
             this.CalculatedPathWeight = -1;
         }
 
-        public int Weight { get; private set; }
+        public long Weight { get; private set; }
 
-        public int CalculatedPathWeight { get; set; }
+        public long CalculatedPathWeight { get; set; }
 
         public bool IsCalculated
         {
